backend/foodgram/requirements.txt
Django==3.2
django-filter==21.1
djangorestframework==3.14.0
djoser==2.1.0
python-decouple==3.5
drf-extra-fields==3.2.1
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.7